refactor: WeatherStatus 패키지 변경
