+++ app/src/main/java/com/example/seoulgonggong/ui/uimodel/UiWeather.kt
@@ -1,7 +1,5 @@
 package com.example.seoulgonggong.ui.uimodel
 
-import com.example.seoulgonggong.domain.model.WeatherStatus
-
 data class UiWeather(
     val temperature: Int,
     val weatherStatus: WeatherStatus,
