+++ app/src/main/java/com/example/seoulgonggong/ui/uimodel/mapper/WeatherMapper.kt
@@ -1,8 +1,8 @@
 package com.example.seoulgonggong.ui.uimodel.mapper
 
 import com.example.seoulgonggong.domain.model.Weather
-import com.example.seoulgonggong.domain.model.WeatherStatus
 import com.example.seoulgonggong.ui.uimodel.UiWeather
+import com.example.seoulgonggong.ui.uimodel.WeatherStatus
 
 fun Weather.toUi(): UiWeather {
     return UiWeather(
